fix : 검색 API 변경
/slangs/search?keyword=신조어 -> /slangs?keyword=신조어
